refactor and comment
