+++ app/src/main/java/com/example/tiptime/MainActivity.kt
@@ -15,51 +15,17 @@
  */
 package com.example.tiptime
 
-import android.graphics.drawable.Icon
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
 import androidx.annotation.VisibleForTesting
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawingPadding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
+import com.example.tiptime.screens.TipCalculatorScreen
 import com.example.tiptime.ui.theme.TipTimeTheme
 import java.text.NumberFormat
 
@@ -72,131 +38,13 @@ class MainActivity : ComponentActivity() {
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                 ) {
-                    TipTimeLayout()
+                    TipCalculatorScreen()
                 }
             }
         }
     }
 }
 
-@Composable
-fun TipTimeLayout() {
-    var amountInput by rememberSaveable {
-        mutableStateOf("")
-    }
-
-    var tipInput by rememberSaveable {
-        mutableStateOf("")
-    }
-
-    var roundUp by rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    val amount = amountInput.toDoubleOrNull() ?: 0.0
-    val tipPercent = tipInput.toDoubleOrNull() ?: 0.0
-
-    val tip = calculateTip(amount = amount, tipPercent = tipPercent, roundup = roundUp)
-
-    Column(
-        modifier = Modifier
-            .statusBarsPadding()
-            .padding(horizontal = 40.dp)
-            .safeDrawingPadding()
-            .verticalScroll(rememberScrollState()),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = stringResource(R.string.calculate_tip),
-            modifier = Modifier
-                .padding(bottom = 16.dp, top = 40.dp)
-                .align(alignment = Alignment.Start)
-        )
-        EditNumberField(
-            label = R.string.bill_amount,
-            value = amountInput,
-            leadingIcon = R.drawable.money,
-            onValueChange = { newState -> amountInput = newState },
-            keyboardOptions = KeyboardOptions.Default.copy(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Next),
-            modifier = Modifier
-                .padding(bottom = 32.dp)
-                .fillMaxWidth()
-            ,
-        )
-        EditNumberField(
-            label = R.string.how_was_the_service,
-            value = tipInput,
-            leadingIcon = R.drawable.percent,
-            onValueChange = { tipInput = it },
-            keyboardOptions = KeyboardOptions.Default.copy(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Done),
-            modifier = Modifier
-                .padding(bottom = 32.dp)
-                .fillMaxWidth()
-            ,
-        )
-        RoundTheTipRow(
-            roundUp = roundUp,
-            onRoundUpChange = { roundUp = it; Log.d("switch","$roundUp") },
-            modifier = Modifier.padding(bottom = 32.dp),
-        )
-        Text(
-            text = stringResource(R.string.tip_amount, tip),
-            style = MaterialTheme.typography.displaySmall
-        )
-        Spacer(modifier = Modifier.height(150.dp))
-    }
-}
-
-@Composable
-fun EditNumberField(
-    @StringRes label: Int,
-    @DrawableRes leadingIcon: Int,
-    value: String,
-    onValueChange: (String)-> Unit,
-    keyboardOptions: KeyboardOptions,
-    modifier: Modifier = Modifier,
-) {
-    TextField(
-        value = value,
-        onValueChange = onValueChange,
-        leadingIcon = { androidx.compose.material3.Icon(painter = painterResource(leadingIcon), null) },
-        label = {Text(stringResource(label))},
-        singleLine = true,
-        keyboardOptions = keyboardOptions,
-        modifier = modifier,
-    )
-}
-
-@Composable
-fun RoundTheTipRow(
-    roundUp: Boolean,
-    onRoundUpChange: (Boolean)-> Unit,
-    modifier: Modifier = Modifier
-){
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .size(48.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            textAlign = TextAlign.Left,
-            text = stringResource(id = R.string.round_up_tip)
-        )
-        Switch(
-            modifier = modifier
-                .fillMaxWidth()
-                .wrapContentWidth(Alignment.End)
-                .testTag("RoundUpSwitch"),
-            checked = roundUp,
-            onCheckedChange = onRoundUpChange)
-    }
-}
 /**
  * Calculates the tip based on the user input and format the tip amount
  * according to the local currency.
@@ -219,6 +67,6 @@ internal fun calculateTip(
 @Composable
 fun TipTimeLayoutPreview() {
     TipTimeTheme {
-        TipTimeLayout()
+        TipCalculatorScreen()
     }
 }
